+++ src/BlockchainReader/Bitcoin.test.ts
@@ -1,0 +1,14 @@
+import { describe } from 'riteway'
+
+import { PREFIX_POET, PREFIX_BARD } from 'Helpers/Bitcoin'
+
+import { blockToPoetAnchors } from './Bitcoin'
+
+describe('Bitcoin', async (should: any) => {
+  const { assert } = should('')
+
+  {
+
+  }
+
+})
