+++ tests/unit/index.ts
@@ -2,6 +2,7 @@ import '../../src/API/Middlewares/RequestValidationMiddleware.test'
 import '../../src/API/Middlewares/RequestValidationMiddleware.test.ts'
 import '../../src/API/Router.test.ts'
 import '../../src/API/WorkController.test.ts'
+import '../../src/BlockchainReader/Bitcoin.test'
 import '../../src/BlockchainReader/BlockchainReaderService.test.ts'
 import '../../src/BlockchainReader/ClaimController.test.ts'
 import '../../src/BlockchainWriter/ClaimController.test.ts'
@@ -18,7 +19,6 @@ import '../../src/View/WorkController.test.ts'
 import '../../src/Extensions/Array.test.ts'
 import '../../src/Extensions/Error.test.ts'
 import '../../src/Extensions/Promise.test.ts'
-import '../../src/BatchReader/Bitcoin.test'
 import '../../src/Helpers/FetchError.test.ts'
 import '../../src/Helpers/Logging.test.ts'
